==> 1.01 Data Sourcing - Web Scraping (BeautifulSoup).ipynb <==
## Import library

import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from time import sleep
import datetime

## Web scraping

current_date=datetime.datetime(year=2021, month=11,day=1)
driver = webdriver.Chrome(ChromeDriverManager().install())
driver.get('https://www.wunderground.com/history/daily/au/melbourne/YMMB/date/'+str(current_date.year)+'-'+str(current_date.month)+'-'+str(current_date.day))
sleep(3)
html = driver.page_source
soup = BeautifulSoup(html,'lxml')
driver.close()
tables = soup.find_all('table')
j=0
k=0
DF=pd.DataFrame(index=np.arange(24), columns=['time','temperature','dew point','humidity','wind',
                                          'wind speed', 'wind gust','pressure','precip.','condition'])
for i in range(100):
    try:
        if j ==0:
            DF.iloc[0,0] =soup.find_all('tr')[i]('td')[0].text
        if j ==1:
            for l in range(10):
                DF.iloc[k,l] =soup.find_all('tr')[i]('td')[l].text
            k=k+1
    except IndexError:
        j=j+1
DF=DF.dropna(axis=0,how='all') 
DF.to_csv(str(current_date.year)+'-'+str(current_date.month)+'-'+str(current_date.day)+'.csv', index=False, encoding = 'utf-8-sig')
