==> 2.01 Data cleaning (Numpy, Pandas).ipynb <==
## Read in Data

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

%matplotlib inline

titanic_data = pd.read_csv('titanic_data.csv')
titanic_data.head(10)

## Check NA values

titanic_data.info()

#check NA values
titanic_data.isnull().sum()

## Clean continuous variables

#### Fill missing for `Age`

titanic_data['Age'].describe()
#titanic_data['Age'].sum()
#titanic_data['Age'].median()
#titanic_data['Age'].mean()
#titanic_data['Age'].max()
#titanic_data['Age'].min()
#titanic_data['Age'].std()
#titanic_data['Age'].var()

#fill missing age with average
titanic_data['Age'].fillna(titanic_data['Age'].mean(), inplace=True)
titanic_data.head(10)

#### Combine `SibSp` & `Parch`

titanic_data['SibSp'].describe()

titanic_data['SibSp'].value_counts()

titanic_data['Parch'].describe()

titanic_data['Parch'].value_counts()

titanic_data['Family_size'] = titanic_data['SibSp'] + titanic_data['Parch']
titanic_data.head(10)

## Clean categorical variables

#### Get `Title` from `Name`

titanic_data['Title'] = titanic_data['Name'].str.split(',').apply(lambda x:x[1].split('.')[0])
titanic_data.head(10)

titanic_data['Title'].describe()

titanic_data['Title'].value_counts()

#### Map `Title`

Title_Dictionary = {
    " Mr":"Mr",
    " Miss":"Miss",
    " Mrs":"Mrs",
    " Master":"Master",
    " Dr":"Officer",
    " Rev":"Officer",
    " Mlle":"Miss",
    " Col":"Officer",
    " Major":"Officer",
    " Ms":"Mrs",
    " Lady":"Royalty",
    " Capt":"Officer",
    " the Countess":"Royalty",
    " Don":"Royalty",
    " Mme":"Mrs",
    " Jonkheer":"Royalty",
    " Sir":"Royalty"
}
titanic_data['Title'] = titanic_data['Title'].map(Title_Dictionary)
titanic_data.head(10)

titanic_data['Title'].value_counts()

#### One-hot encode with `Title`

Title_Df =pd.get_dummies(titanic_data['Title'],prefix='Title')
titanic_data = pd.concat([titanic_data, Title_Df], axis=1)
titanic_data.head(10)

#### Map `Sex`

titanic_data['Sex'].value_counts()

sex_mapDict={
    'male':1,
    'female':0
}
titanic_data['Sex'] = titanic_data['Sex'].map(sex_mapDict)
titanic_data.head(10)

#### Map `Cabin`

titanic_data['Cabin'] = titanic_data['Cabin'].str[0:1]
titanic_data.head(10)

titanic_data['Cabin'].value_counts()

#### One-hot encode with `Cabin`

Cabin_Df =pd.get_dummies(titanic_data['Cabin'],prefix='Cabin')
titanic_data = pd.concat([titanic_data, Cabin_Df], axis=1)
titanic_data.head(10)

#### One-hot encode with `Embarked`

titanic_data['Embarked'].value_counts()

embarked_Df =pd.get_dummies(titanic_data['Embarked'],prefix='Embarked')
titanic_data = pd.concat([titanic_data, embarked_Df], axis=1)
titanic_data.head(10)

## Drop unnecessary variables

titanic_data.drop(['PassengerId', 'Name', 'SibSp', 'Parch','Ticket','Cabin','Embarked','Title'], axis=1, inplace=True)
titanic_data.head(10)

## Write out cleaned data

titanic_data.to_csv('cleaned_titanic_data.csv', index=False)

## Split data into train data and test data

features = titanic_data.drop('Survived', axis=1)
labels = titanic_data['Survived']

X_train, X_test, y_train, y_test = train_test_split(features, labels, test_size=0.2, random_state=42)

X_train.to_csv('train_features.csv', index=False)
X_test.to_csv('test_features.csv', index=False)

y_train.to_csv('train_labels.csv', index=False)
y_test.to_csv('test_labels.csv', index=False)
