==> 4.04 Regression Analysis - Random Forest (sklearn).ipynb <==
### Import library

import pandas as pd                   
import matplotlib.pyplot as plt        
import seaborn as sns           
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV

### Prepare Data

train_features = pd.read_csv('train_features.csv')
train_labels = pd.read_csv('train_labels.csv')
train_labels = train_labels.Survived

test_features = pd.read_csv('test_features.csv')
test_labels = pd.read_csv('test_labels.csv')
test_labels = test_labels.Survived
test_features

### Train model

def print_results(results):
    print('BEST PARAMS: {}\n'.format(results.best_params_))

    means = results.cv_results_['mean_test_score']
    stds = results.cv_results_['std_test_score']
    for mean, std, params in zip(means, stds, results.cv_results_['params']):
        print('{} (+/-{}) for {}'.format(round(mean, 3), round(std * 2, 3), params))
        
rf = RandomForestRegressor()
parameters = {
    'n_estimators': [5, 50, 250],
    'max_depth': [2, 4, 8, 16, 32, None]
}

cv = GridSearchCV(rf, parameters, cv=5)
cv.fit(train_features[['Age','Pclass']].values, train_features['Fare'].values)

print_results(cv)

best_rf=cv.best_estimator_

### Test Model

print(
    'Score on testing data: ' 
    + str("{:.2%}".format(best_rf.score(test_features[['Age','Pclass']].values, test_features['Fare'].values))))

best_rf.predict(test_features[['Age','Pclass']].values)
