==> 8.02 Statistics - Error Analysis (Sklearn).ipynb <==
### Import Library

import pandas as pd     
import numpy as np
import seaborn as sns
from scipy.stats import pearsonr
from scipy.stats import spearmanr
from sklearn.metrics import mean_squared_error as MSE
from sklearn.metrics import mean_absolute_error as MAE

### Prepare Data

test_features = pd.read_csv('test_features.csv')
test_features = test_features[['Age','Fare']]
test_features.head(10)

### Calculate RMSE

MSE = MSE(test_features['Age'], test_features['Fare'])
RMSE = np.sqrt(MSE)
RMSE

### Calculate RMSE

MAE = MAE(test_features['Age'], test_features['Fare'])
MAE
