==> 8.03 Statistics - Test of Independence (Scipy).ipynb <==
### Import Library

import pandas as pd     
import numpy as np
import seaborn as sns
from scipy.stats import ttest_rel
from scipy.stats import ttest_ind
from scipy.stats import wilcoxon
from scipy.stats import mannwhitneyu

### Prepare Data

test_features = pd.read_csv('test_features.csv')
test_features = test_features[['Age','Fare']]
test_features.head(10)

### Conduct Paired Samples t Test

stat, p = ttest_rel(test_features['Age'], test_features['Fare'])
p

### Conduct Independent t Test

stat, p = ttest_ind(test_features['Age'], test_features['Fare'])
p

### Conduct Pairwise Wilcoxon Rank Sum Test

stat, p = wilcoxon(test_features['Age'], test_features['Fare'])
p

### Conduct MannWhitney U Test

stat, p = mannwhitneyu(test_features['Age'], test_features['Fare'])
p
