==> 9.01 Textual data analaysis (nltk).ipynb <==
### Import Library

import re  
import nltk 
import matplotlib.pyplot as plt  
import pandas as pd  
from nltk.tokenize import word_tokenize
from nltk.corpus import opinion_lexicon

### Prepare Data

df = pd.read_csv('LittleWomen.txt',sep='\t')\
    .dropna()\
    .drop('gutenberg_id', 1)

df.head(10)

def clean_text(text):
    text = text.lower() 
    text = text.replace("'", '')
    text = re.sub(r'[^\w]', ' ', text) 
    text = re.sub(r'\s+', ' ', text)
    text = text.strip()
    return text

df['text'] = df['text'].map(clean_text) 
df['text'] = df['text'].map(word_tokenize) 

df = df.text.explode().to_frame('token')
df.head()

### Summarise the Sentiment Words

sentiment_lexicon = {
    **{w: 'positive' for w in opinion_lexicon.positive()},
    **{w: 'negative' for w in opinion_lexicon.negative()}
}

df['sentiment'] = df['token'].map(sentiment_lexicon)
df = df[~df.sentiment.isna()]
df.head(10)

summary_df = df.sentiment.value_counts().to_frame('n')
summary_df['prop'] = summary_df['n'] / summary_df.n.sum()

summary_df.round(3)
